Rearrange task outcome and caching tests (#1323)
* A new FilteringCachingTest

* Tweak contains checks

* Use assertCompositeExecutions

* Remove SUCCESS check for canRegisterCustomShadowJarTask

* Remove FAILED check for failBuildIfProcessingBadJar

* Remove FAILED check for honorDuplicatesStrategyWithThrowing

* Remove configurationCachingOfConfigurationsIsUpToDate as no inputs changed

* Fix style

* Reduce all assertions
