+++ src/functionalTest/kotlin/com/github/jengelman/gradle/plugins/shadow/FilteringTest.kt
@@ -4,9 +4,7 @@ import assertk.assertThat
 import com.github.jengelman.gradle.plugins.shadow.util.containsEntries
 import com.github.jengelman.gradle.plugins.shadow.util.doesNotContainEntries
 import kotlin.io.path.appendText
-import kotlin.io.path.readText
 import kotlin.io.path.writeText
-import org.gradle.testkit.runner.TaskOutcome.SUCCESS
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -92,57 +90,6 @@ class FilteringTest : BasePluginTest() {
   }
 
   @Test
-  fun dependencyExclusionsAffectUpToDateCheck() {
-    dependOnAndExcludeArtifactD()
-
-    run(shadowJarTask)
-
-    commonAssertions()
-
-    val replaced = projectScriptPath.readText()
-      .replace("exclude(dependency('my:d:1.0'))", "exclude(dependency('my:c:1.0'))")
-    projectScriptPath.writeText(replaced)
-    val result = run(shadowJarTask)
-
-    assertThat(result).taskOutcomeEquals(shadowJarTask, SUCCESS)
-    assertThat(outputShadowJar).useAll {
-      val entries = entriesInAB + "d.properties"
-      containsEntries(*entries)
-      doesNotContainEntries(
-        "c.properties",
-      )
-    }
-  }
-
-  @Test
-  fun projectExclusionsAffectUpToDateCheck() {
-    dependOnAndExcludeArtifactD()
-
-    run(shadowJarTask)
-
-    commonAssertions()
-
-    val replaced = projectScriptPath.readText()
-      .replace("exclude(dependency('my:d:1.0'))", "exclude 'a.properties'")
-    projectScriptPath.writeText(replaced)
-
-    val result = run(shadowJarTask)
-
-    assertThat(result).taskOutcomeEquals(shadowJarTask, SUCCESS)
-    assertThat(outputShadowJar).useAll {
-      containsEntries(
-        "a2.properties",
-        "b.properties",
-        "c.properties",
-        "d.properties",
-      )
-      doesNotContainEntries(
-        "a.properties",
-      )
-    }
-  }
-
-  @Test
   fun includeDependencyAndExcludeOthers() {
     projectScriptPath.appendText(
       """
@@ -170,8 +117,10 @@ class FilteringTest : BasePluginTest() {
         "d.properties",
         "my/Passed.class",
       )
-      val entries = entriesInAB + "c.properties"
-      doesNotContainEntries(*entries)
+      doesNotContainEntries(
+        *entriesInAB,
+        "c.properties",
+      )
     }
   }
 
@@ -273,8 +222,10 @@ class FilteringTest : BasePluginTest() {
 
   private fun commonAssertions() {
     assertThat(outputShadowJar).useAll {
-      val entries = entriesInAB + "c.properties"
-      containsEntries(*entries)
+      containsEntries(
+        *entriesInAB,
+        "c.properties",
+      )
       doesNotContainEntries(
         "d.properties",
       )
